chore(react-aria): restructure folder organization
