+++ change/@fluentui-react-aria-01c1906f-a6a1-431d-b4a2-e8f0310c26eb.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: restructure folder organization",
+  "packageName": "@fluentui/react-aria",
+  "email": "bernardo.sunderhus@gmail.com",
+  "dependentChangeType": "none"
+}
